[rbp] canged omx jpeg decoding
image decoding is not done in in an egl image anymore.
the reason for this are the changes in the background loader using CTexture.

dissadvantage :

decoding and rendering needs a tick longer.

advantage :

no risk to run out of GPU memory on to many loaded textures.
thumbnailing should also take advantage of the decoder.